update for local registry

+++ main.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 
+import os
 import shutil
 from typing import Tuple
 
@@ -14,7 +15,6 @@ import util
 
 DATA_DIR = "data"
 LOG_DIR = "log"
-TEMP_DIR = "tmp"
 CONFIG = "config.yaml"
 PREFETCH_FILE_LIST = "algorithm/out.txt"
 
@@ -93,7 +93,10 @@ def start_bench(cfg: dict, image: str):
 
 if __name__ == "__main__":
     util.clean_env()
-    main()
-    shutil.rmtree(TEMP_DIR)
-    shutil.rmtree(DATA_DIR)
-    shutil.rmtree(LOG_DIR)
+    try:
+        main()
+    except Exception:
+        print(Exception)
+        util.reload_nydus()
+    if os.path.exists(LOG_DIR):
+        shutil.rmtree(LOG_DIR)
